+++ packages/dso-toolkit/fractal/collator.js
@@ -3,9 +3,9 @@ module.exports = {
     `<!-- Start: @${item.handle} -->\n` +
     (item.context.__title || item.context.__explanation || item.context.__switches
       ? `
-        <div class="dso-example-wrapper">
+        <div class="toolkit-example-wrapper">
           ${item.context.__title || item.context.__explanation
-            ? `<div class="dso-example-text">
+            ? `<div class="toolkit-example-text">
                 ${item.context.__title ? `<h2>${item.context.__title}</h2>` : ''}
                 ${item.context.__explanation
                   ? [].concat(item.context.__explanation).map(e => `<p>${e}</p>`).join('\n')
@@ -15,7 +15,7 @@ module.exports = {
           ${item.context.__switches && item.context.__switches.length > 0
             ? item.context.__switches.map(s => `<input type="checkbox" id="${s.name}-switch" /><label for="${s.name}-switch">${s.label}</label>`)
             : ''}
-          <div class="dso-example">
+          <div class="toolkit-example">
             ${markup}
           </div>
         </div>`