+++ packages/dso-toolkit/theme/index.js
@@ -198,13 +198,13 @@ module.exports = function (options) {
       });
 
       $hydrated('body')
-        .find('.container')
+        .find('.toolkit-example-container')
           .prepend('<h2>Markup component preview</h2>')
         .end()
         .prepend('<hr id="custom-elements-raw">')
         .prepend(raw)
-        .find('.container:first-child')
-          .prepend('<h2>Web Component preview</h2>');;
+        .find('.toolkit-example-container:first-child')
+          .prepend('<h2>Web Component preview</h2>');
 
       const result = $hydrated.html();